pyproject.toml
[project]
name = "ai-portfolio-be"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "openai>=1.68.2",
    "python-dotenv>=1.0.1",
    "pypdf>=5.4.0"
]